requirements.txt
openai
heroku
flask
gunicorn
wtforms
Flask-WTF
Bootstrap-Flask
waitress
wheel
python-dotenv
Flask-CKEditor